refactor: make node._from_rest_object logic reusable
