+++ sdk/ml/azure-ai-ml/azure/ai/ml/_internal/entities/command.py
@@ -92,11 +92,13 @@ class Command(InternalBaseNode):
         return rest_obj
 
     @classmethod
-    def _rest_object_to_init_params(cls, obj):
-        obj = InternalBaseNode._rest_object_to_init_params(obj)
+    def _from_rest_object_to_init_params(cls, obj):
+        obj = InternalBaseNode._from_rest_object_to_init_params(obj)
 
-        # resources
         if "resources" in obj and obj["resources"]:
+            from azure.ai.ml._restclient.v2022_10_01_preview.models import \
+                JobResourceConfiguration as RestJobResourceConfiguration
+            from azure.ai.ml.entities._job.job_resource_configuration import JobResourceConfiguration
             resources = RestJobResourceConfiguration.from_dict(obj["resources"])
             obj["resources"] = JobResourceConfiguration._from_rest_object(resources)
 
@@ -149,16 +151,6 @@ class Distributed(Command):
         from .._schema.command import DistributedSchema
 
         return DistributedSchema(context=context)
-
-    @classmethod
-    def _rest_object_to_init_params(cls, obj: dict):
-        obj = Command._rest_object_to_init_params(obj)
-
-        # distribution
-        if "distribution" in obj and obj["distribution"]:
-            obj["distribution"] = DistributionConfiguration._from_rest_object(obj["distribution"])
-
-        return obj
 
     @classmethod
     def _picked_fields_from_dict_to_rest_object(cls) -> List[str]:
